dynamic parent class wip
