+++ Gemfile
@@ -9,4 +9,5 @@ group :development, :test do
   gem 'simplecov'
   gem 'timecop'
   gem 'waterdrop'
+  gem 'activesupport'
 end
