+++ spec/spec_helper.rb
@@ -9,6 +9,7 @@ $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
   timecop
   fiddle
   tempfile
+  active_support/hash_with_indifferent_access
 ].each do |lib|
   require lib
 end
